Moved all unnecessary external declarations into cpp file
